domainbed/requirements.txt
numpy==1.20.3
wilds==1.2.2
imageio==2.9.0
gdown==3.13.0
torch==2.0.0
tqdm==4.62.2
backpack==0.1
parameterized==0.8.1
Pillow==8.3.2
fsspec
s3fs
fair-esm>=2.0.0
addict
